refactory: int -> int64_t

+++ varlisp/detail/io.hpp
@@ -9,11 +9,11 @@
 
 namespace varlisp {
 namespace detail {
-inline std::string readline(int fd)
+inline std::string readline(int64_t fd)
 {
     std::string line;
     char ch = '\0';
-    int ec = 0;
+    int64_t ec = 0;
     while ((ec = ::read(fd, &ch, 1)) != -1) {
         if (ec == 0) {
             return line;
@@ -25,16 +25,16 @@ inline std::string readline(int fd)
     }
     return line;
 }
-inline int readchar(int fd)
+inline int64_t readchar(int64_t fd)
 {
     char buf[6];
-    int ec = ::read(fd, buf, 1);
+    int64_t ec = ::read(fd, buf, 1);
     if (ec == -1) {
         return -1;
     }
-    int len = sss::util::utf8::next_length(buf, buf + 1);
+    int64_t len = sss::util::utf8::next_length(buf, buf + 1);
     if (len == 1) {
-        return int(buf[0]);
+        return int64_t(buf[0]);
     }
     else {
         ec = ::read(fd, buf + 1, len - 1);
@@ -45,15 +45,15 @@ inline int readchar(int fd)
         if (ch.second != len) {
             return -1;
         }
-        return int(ch.first);
+        return int64_t(ch.first);
     }
 }
 
-inline int writechar(int fd, int ch)
+inline int64_t writechar(int64_t fd, int64_t ch)
 {
     char buf[6];
     if (ch && ch != -1) {
-        int len = sss::util::utf8::get_ucs_code_length(ch);
+        int64_t len = sss::util::utf8::get_ucs_code_length(ch);
         sss::util::utf8::dumpout2utf8(&ch, &ch + 1, buf);
         // 返回值包括-1状态！
         return ::write(fd, buf, len);;
@@ -61,7 +61,7 @@ inline int writechar(int fd, int ch)
     return 0;
 }
 
-inline int writestring(int fd, sss::string_view sv)
+inline int64_t writestring(int64_t fd, sss::string_view sv)
 {
     return ::write(fd, sv.data(), sv.size());
 }
